[composables] Extract useTaskStatusStyle composable
Extract duplicated task status color logic from ComboboxStatus and
ValidationTag into a shared useTaskStatusStyle composable.
